+++ R/vcov.cirls.R
@@ -29,15 +29,14 @@ vcov.cirls <- function(object, complete = TRUE, nsim = 1000, constrained = TRUE,
     # simulate from truncated multivariate normal
     seed <- if ("seed" %in% names(dots)) dots$seed else NULL
     simures <- simulCoef(object, nsim = nsim, complete = complete,
-      seed = seed)
+      seed = seed, constrained = TRUE)
 
     # Compute empirical variance
     v <- stats::var(simures)
   } else {
 
     # If constrained = FALSE return the unmodified vcov
-    v <- stats::.vcov.aliased(aliased, stats::summary.glm(object)$cov.scaled,
-      complete = complete)
+    v <- stats::vcov(uncons(object), complete = complete)
 
   }
   # Return